requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pgvector
openai
pydantic
sentence-transformers
pandas